refactor(sentence): moved findKeys-functionality to util findKeysInPlaceholder + cleanup
