+++ src/sentence/utils/format.ts
