--- src/sentence/utils/index.ts
@@ -1,17 +1,0 @@
-export {
-  getTotalWeightOfEntries,
-  mapToWeightedEntryArray,
-  pickRandomEntryByWeight,
-} from './array';
-
-export {
-  articleAndPluralize,
-  capitalize,
-  isVowel,
-} from './format';
-
-export {
-  escapeSpecialCharacters,
-  findPlaceholdersByNotation,
-  parsePlaceholderNotation,
-} from './placeholders';
