+++ src/sentence/utils/placeholder.ts
@@ -19,3 +19,13 @@ export const findPlaceholdersByNotation = (template: StringResolvable, placehold
   const regex = new RegExp(`([${escapeSpecialCharacters(start)}]+(\\s*([a-z-0-9])*,?\\s*)*[${escapeSpecialCharacters(end)}]+)`, 'gi');
   return (typeof template === 'string') ? template.match(regex) : template().match(regex);
 };
+
+/**
+ * Returns keys found in the given placeholder
+ * @param {string} placeholder
+ * '{a-adjective, a-curse, verb}' => ['a-adjective', 'a-curse', 'verb']
+ */
+export const findKeysInPlaceholder = (placeholder: string, placeholderNotation: PlaceholderNotation): string[] => {
+  const { start, end } = placeholderNotation;
+  return placeholder.replace(new RegExp(`${escapeSpecialCharacters(start)}|${escapeSpecialCharacters(end)}|\\s`, 'g'), '').split(',');
+};
